Fixed tests, Added bogus for data generation and Fixed frontend grid pagination

+++ api/src/WebApp.Template.Application/Data/DbEntities/Location.cs
@@ -9,8 +9,8 @@ public class Location
 {
     public string Id { get; init; }
     public string Name { get; init; }
-    public decimal Longitude { get; init; }
-    public decimal Latitude { get; init; }
+    public double Longitude { get; init; }
+    public double Latitude { get; init; }
 
     /* Navigation properties */
     public ICollection<Plant> Plants { get; init; }
@@ -18,7 +18,7 @@ public class Location
     /* Private constructor for EF Core */
     private Location() { }
 
-    public static Location CreateLocation(string id, string name, decimal longitude, decimal latitude)
+    public static Location CreateLocation(string id, string name, double longitude, double latitude)
     {
         var newLocation = new Location
         {
